refactor: Move condition classes to separate folder
