+++ test/unit/http/ldp/GetOperationHandler.test.ts
@@ -4,8 +4,8 @@ import type { Operation } from '../../../../src/http/Operation';
 import { BasicRepresentation } from '../../../../src/http/representation/BasicRepresentation';
 import type { Representation } from '../../../../src/http/representation/Representation';
 import { RepresentationMetadata } from '../../../../src/http/representation/RepresentationMetadata';
-import { BasicConditions } from '../../../../src/storage/BasicConditions';
-import { getETag } from '../../../../src/storage/Conditions';
+import { BasicConditions } from '../../../../src/storage/conditions/BasicConditions';
+import { getETag } from '../../../../src/storage/conditions/Conditions';
 import type { ResourceStore } from '../../../../src/storage/ResourceStore';
 import { NotImplementedHttpError } from '../../../../src/util/errors/NotImplementedHttpError';
 import { NotModifiedHttpError } from '../../../../src/util/errors/NotModifiedHttpError';
